# gantt-backend/api/api-spec.yaml
openapi: 3.0.0
info:
  title: Authorization API
  version: 1.0.0
security:
  - bearerAuth: []
paths:
  /api/check-email:
    post:
      security: [ ]
      summary: "Проверить существование email"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
              required:
                - email
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  exists:
                    type: boolean
                  message:
                    type: string
  /api/register:
    post:
      security: [ ]
      summary: "Регистрация"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                nickname:
                  type: string
                password:
                  type: string
              required:
                - email
                - nickname
                - password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '409':
          $ref: '#/components/responses/Conflict'
  /api/login:
    post:
      security: [ ]
      summary: "Авторизация"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                password:
                  type: string
              required:
                - email
                - password
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/user/{user_id}:
    get:
      summary: "Получить информацию о пользователе"
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
                  teams:
                    type: array
                    items:
                      $ref: '#/components/schemas/TeamWithId'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      summary: "Частично обновить данные о пользователе"
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                user:
                  type: object
                  properties:
                    mail:
                      type: string
                      nullable: true
                    nickname:
                      type: string
                      nullable: true
                    oldPassword:
                      type: string
                    newPassword:
                      type: string
                      nullable: true
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  user:
                    $ref: '#/components/schemas/User'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'

    delete:
      summary: "Удалить пользователя"
      parameters:
        - $ref: '#/components/parameters/UserIdPath'
      responses:
        '204':
          $ref: '#/components/responses/NoContent'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  schemas:
    AuthResponse:
      type: object
      properties:
        access_token:
          type: string
        token_type:
          type: string
    User:
      type: object
      properties:
        id:
          type: integer
        mail:
          type: string
        nickname:
          type: string
...
    TeamWithId:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
...
  parameters:
    UserIdPath:
      name: user_id
      in: path
      required: true
      schema:
        type: integer
...
  responses:
    NoContent:
      description: No content
    BadRequest:
      description: Bad request
    Unauthorized:
      description: Unauthorized
    Forbidden:
      description: Forbidden
    NotFound:
      description: Not found
    Conflict:
      description: Conflict